♻️ 코드 수정 : LectureFacade 패키지명 수정
